Wrap search calls with a circuit breaker
Make sure we give some time for the service to recover if we enter
a situation where the error rate goes too high.

This patch introduces a circuit breaker for the calls that go
through the parallel scheduler so that we stop accepting queries for
a while (60s by default) if the error rate gets too high (50% by
default).

A health check (under /actuator/health/circuibreaker) has been added
and metrics are exposed via /actuator/prometheus and
/actuator/metrics/resilience4j.*