Add basic pagination support
This patch implements pagination by means of holding all items
in the result until the desired offset is reached. This results
in a functionality that works well for the first pages but
grows progressivelly more expensive the further we are from
the top results.

A better theoretical approach would be to hold a reference to
a previous bottom item (as in: its Id and Score (and more if
sharding)) and use it as the threshold so that instead of keeping
every single ranked item in memory we can discard everything
that scores higher than it (meaning: memory usage becomes a
function of the page size, not page size + offset). Lucene
provides a functionality similar to this via searchAfter
but it isn't super trivial to use because the sorting API is
untyped so we need to be more careful creating the reference
point instance.

This patchs implementation works ok, especially with <1M documents
to be looking at, and it's easier to roll since it's just a one
way communication channel (front-end just needs to specify the
page number- searchAfter requires a previous search result to
know the pagination parameters). I think I'll keep it.