Only run the breaker in the searching threads
Now we only fire completion/timeout events based on the time
to search, not for the whole async pipeline.

Search is the only expensive thing that happens, after that it's
pretty much just IO so I think it's better to just flush it
down the pipe instead of aborting the whole thing.