Somewhat stricter static resource matching
I'm starting to understand how this PathPattern is used by the resource
handlers, but it's too weird to me that this:

    .addResourceHandler("/css/main.css")

Does not work irrespective of what I tried for resource location. If
I want to be super strict, the closest I got was:

    .addResourceHandler("/css/{filename:main.css}")

But that's too silly. So it looks like it requires a capture, either
explicit or implicit. Anyway, not worth figuring it out properly
right now ¯\_(ツ)_/¯