Handle broken /recipe/ URIs correctly
Before this patch, handling a uri with a bad recipeId part such as:

    /recipe/valid-slug/something-that-doesnt-parse-as-number

Would yield a 5xx-type error page because the recipeId field would
fail to be parsed as a number with `NumberFormatException` and the
ExceptionHandler had no knowledge of it.

Tests are added to guarantee that the endpoint behaves as expected.