Add some validation for sortOptions and dietFilterOptions
This is not a decent solution, but updating SortOption or Diet in
cerberus doesn't trigger a compile-time breakage in this project
(but it breaks in production and in tests because of runtime
checks).

This patch makes the code break during a `static{}` block
initialization if we're using an invalid value.