SidebarComponent: Add support for counts
While this works, I'm not happy with the coupling with cerberus and
when I was reviewing this for committing I had an idea:

Now that cerberus only does taxonomy facets it's easy to expose an
api that allows the clients to configure which facets to index and
retrieve. Something like:

    new Indexer.Builder()
            .dataDirectory(Path.of("tmp/dir"))
            .withFacets(recipe -> {
                ...
                if (recipe.ingredients().size() <= 5) {
                    facets.addFacet("num_ingredient", "0,5");
                }
                ...
                return facets;
            })
            .build();

Which means many good things:

1. Cerberus can stop caring about facet and label names,
   it would all be driven by the client

2. Controlling the facet data means that this mess of sharing string
   interfaces (read this patche's TODOs) isn't a problem at all:
   the SidebarComponent (Or whatever comes up from it) can drive
   everything and control all constants.

3. The Loader will finally be moved to this project, where it
   belongs. This will force me to actually care about that code.

Anyway, committing because this is work almost complete and perhaps
this `withFacets` idea doesn't fly for some reason I can't think of
right now.