Fully working Typed SidebarComponent
There's still much to be done, but I'm quite happy with the results
of moving the facets to the client. Code is a bit more verbose and
there's still significant work to be done, but I'll be able to sync
up all constants giving me a lot more confidence when changing this
stringly-driven part of the code.