Make Sidebar options typed too
Now all options are typed and addressable and only valid options
are ever indexed. No more Strings and toString() being flung
about as if this was bash or something similar.

I'm abusing enums a bit to achieve what I want, but I'm sure that
sometime in the future we'll get ADTs and we'll all rejoice.

There's still a bit more to be done (mostly moving and renaming
things) and some of the new/smarter functionality needs tests.

Also, shout out to past me for everything at SidebarComponentTest:
the tests saved me a LOT of debugging time, especially while
this patch was being written and things were seriously messy.