Isolate RecipeMetadata* as RecipeMetadataService
This patch introduces RecipeMetadataService as a simple wrapper over
cerberus' RecipeMetadataDatabase so that there's a single entry
point, making it easier to adapt to interface evolutions.

Now ModelView doesn't know how to fetch recipe metadata, instead it
receives it directly or is told how to retrieve the recipes via
a reference to the service.

This incorrectSlugYieldsNotFound and incorrectIdYieldsNotFound tests
were dropped as BootApplicationTest already tests this behaviour.