Answer HEAD requests /, /search and /recipe properly
This patch simply exposes a route for said paths akin to the ones
implemented for GET requests. This is a lot more verbose than
necessary: technically I should be able to dispatch to the same
methods in the router and avoid any unnecessary work, but this
will need a bit work in how I assemble the reactor pipeline.