De-specialize HEAD handling
Drop this idea of a smarter HEAD handling. I don't think it's necessary
at the moment to care so much about useless things being lean. I,
however, quite like the state of the code after laying the groundwork
for pushing the feature, so that stays and I'll revisit this maybe
never.