Compute stats when no flags are given
Also control parallelism via project reactor because it's easier to
pick the scheduler and tell it to shutdown. Not strictly necessary,
but `mvn exec:java` by default waits for all deamon threads to
terminate but `Stream.parallel()` sets up a pool. Using ParallelFlux
makes it trivial to declare which pool to use and thus manually
terminate it on program completion.