Use a fancy SearchPolicy to rewrite too-expensive queries
This patch switches the policy to the new interface and makes it rewrite
queries that are too expensive into a simple MatchAllDocsQuery(). The
approach comes from experimentation:

Terms like "tender", "lemon" and "skillet" each hit about 200k documents
in the index currently. Searching for any combination of these ("tender
lemon", for example), however, is slower than scanning the index
(about 10% less throughput); So our policy now converts a query for
such heavy terms into a simple query that matches all documents.

Ideally this will: (1) almost never happen and (2) notify the user when
it does, but for now all that we do is silently rewrite.

This patch also bumps cerberus to 0.5.0, which is two feature releases
above the previous stable one. 0.4.x released the yet-unused
`findSimilar` (It's a WIP in tablier) searcher API.