Allow arbitrary queries, even empty ones
This patch makes the search &q= parameter optional, so that one can
search without typing anything (or just white-space, or "*") and
browse all recipes. It's useful for answering questions like:

> What is the quickest paleo meal to prepare

This pulls in `cerberus-0.5.3-SNAPSHOT` since the 0.5.2 release
contains logic that does not allow a query to be built without
the `fulltext` field set.