SearchPolicy: Log expensive queries i/o rewriting
This patch simplifies the logic in `MagicQuery.rewrite` (class renamed
to PerformanceInspectorQuery) to- instead of emitting a MatchAllDocsQuery
or throwing- warn when such queries are detected. I have a bit more
confidence to do this now that we cache queries, so abusing my little
server's CPU is slightly harder.

Prior to cerberus-0.6.1 this will end up logging twice due to how the
library handled search queries. Now this should only log one for every
_uncached_ possibly expensive query.