Stop adding url parameters to recipe's infoUrl
Since I stopped using an access.log, these parameters became useless.
They were used as a way to provide attribution:

If the log contains a info url, I can look at the parameters to
reconstruct the search that led to the page thus being able to model
it as an implicit feedback (i.e.: a graph model would say that the
recipe this infoUrl shows is preferred over the topK recipes the search
results contained).

Now I use privacy as a feature: nothing is logged. So my models need
to be anonymous and this idea of pushing parameters around is
useless.