Always preserve filters
This patch points at tablier's interim release so that we can finally
enable a feature where every filter option is preserved.

What this means is that, now, if you search from a Search Result page,
any selected filter will be preserved for the subsequent search. This
is particularly helpful for a use-case I've been needing for a while:

Bookmark-able, reuse-able urls. One may now bookmark a search for
their preferences (say: `-peanut diet:keto sort:total_time`) and now
they can use it directly as if it were a specialized site.