Provide a RecipeMetadataDatabase for BootApplicationTest
Before this patch things were working "accidentally": in order to
instantiate the (mocked) metadata service it was instantiating the
RecipeMetadataDatabase bean declared at BootApplication class, so
if that failed, the tests also did. Now we actually provide a
replacement bean that's guaranteed to not fail because of missing
data/configuration.