Check for cached queries outside the parallel scheduler
This patch makes the SearchService eagerly return the result if found
on cache instead of submitting it to the executor pool. This is being
done to reduce CPU time spent not searching inside the parallel
worker pool.

Before this change, when under load the /search endpoint would behave
in a exclusively synchronous manner, so even a repeatedly hammered
query could starve our cores. Now cached queries never hit the
parallel executors so we can keep servicing real non-cached queries.