Add the current commit id to metric tags
This patch re-enables the git-commit-id plugin execution, sets it up
so that it writes the application.properties:git.head.describe value
and then customizes the meter registry settings to include the new
tag.

The "application" tag introduced on bb361aa5 has been moved to be
defined on code as well, so the common tags configuration only happens
at one place.